WIP: Programatically docking for first time runs

+++ graphics/src/imgui/ImGuiRenderer.cpp
@@ -56,7 +56,7 @@ namespace Graphics {
     void ImGuiRenderer::Render(const WindowProperties& windowProperties) const
     {
         ImGuiIO& io = ImGui::GetIO();
-        io.DisplaySize = ImVec2(static_cast<float>(windowProperties.Resolution.Width), static_cast<float>(windowProperties.Resolution.Height));
+        io.DisplaySize = ImVec2(static_cast<int>(windowProperties.Resolution.Width), static_cast<int>(windowProperties.Resolution.Height));
 
         ImGui::Render();
         ImGui_ImplOpenGL3_RenderDrawData(ImGui::GetDrawData());
